+++ ewm-main-service/src/main/java/ru/practicum/mapper/ParticipationMapper.java
@@ -6,7 +6,7 @@ import ru.practicum.model.participation.enums.ParticipantState;
 import java.time.LocalDateTime;
 
 public class ParticipationMapper {
-    public static ParticipationRequestDto toDto(LocalDateTime created, int eventId, int id, int requester, ParticipantState status) {
+    public static ParticipationRequestDto toDto(LocalDateTime created, Long eventId, int id, int requester, ParticipantState status) {
         return new ParticipationRequestDto(created, eventId, id, requester, status);
     }
 }
