+++ ewm-main-service/src/main/java/ru/practicum/model/participation/dto/ParticipationRequestDto.java
@@ -20,7 +20,7 @@ public class ParticipationRequestDto {
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd HH:mm:ss")
     LocalDateTime created;
 
-    int event;
+    Long event;
 
     int id;
