+++ ewm-main-service/src/main/java/ru/practicum/repository/event/ParticipationRepository.java
@@ -17,7 +17,7 @@ public interface ParticipationRepository extends JpaRepository<Participation, In
 
     List<Participation> findAllByStatusEquals(ParticipantState participantState);
 
-    List<Participation> findAllByEvent_IdAndEvent_Initiator_Id(Integer eventId, Integer userId);
+    List<Participation> findAllByEvent_IdAndEvent_Initiator_Id(Long eventId, Integer userId);
 
     List<Participation> findAllByRequester_Id(Integer userId);
