+++ ewm-main-service/src/main/java/ru/practicum/service/user/UserCompilationServiceImpl.java
@@ -81,7 +81,7 @@ public class UserCompilationServiceImpl implements UserCompilationService {
         return participations.size();
     }
 
-    private long getViews(int eventId) {
+    private long getViews(Long eventId) {
         String uri = "events/" + eventId;
         List<HitResponseDto> viewsList;
         try {
